Saving window state, changed design, new bugs fixed, support parameters for modules, tests

+++ test/edu/mgupi/pass/filters/ParamHelperTest.java
@@ -100,7 +100,7 @@ public class ParamHelperTest {
 		List<Param> paramList = new ArrayList<Param>();
 		paramList.add(new Param("key1", "Ключ 1", TYPES.STRING, "14.55"));
 		paramList.add(new Param("key2", "Ключ 2", TYPES.INT, 6));
-		paramList.add(new Param("key3", "Ключ 3", TYPES.INT, 12.22));
+		paramList.add(new Param("key3", "Ключ 3", TYPES.DOUBLE, 12.22));
 
 		paramList.get(paramList.size() - 1).setValue(44.56);
 
@@ -117,7 +117,7 @@ public class ParamHelperTest {
 		ArrayList<Param> paramList = new ArrayList<Param>();
 		paramList.add(new Param("key1", "Ключ 1", TYPES.STRING, "14.55"));
 		paramList.add(new Param("key2", "Ключ 2", TYPES.INT, 6));
-		paramList.add(new Param("key3", "Ключ 3", TYPES.INT, 12.22));
+		paramList.add(new Param("key3", "Ключ 3", TYPES.DOUBLE, 12.22));
 
 		MyFilter myFilter = new MyFilter(paramList);
 
@@ -145,7 +145,7 @@ public class ParamHelperTest {
 		ArrayList<Param> paramList = new ArrayList<Param>();
 		paramList.add(new Param("key1", "Ключ 1", TYPES.STRING, "14.55"));
 		paramList.add(new Param("key2", "Ключ 2", TYPES.INT, 6));
-		paramList.add(new Param("key3", "Ключ 3", TYPES.INT, 12.22));
+		paramList.add(new Param("key3", "Ключ 3", TYPES.DOUBLE, 12.22));
 
 		MyFilter myFilter = new MyFilter(paramList);
 		assertEquals("{\"key1\":\"14.55\",\"key2\":6,\"key3\":12.22}", ParamHelper.convertParamsToJSON(myFilter));
